COOK-1294: Chef 10.10 doesn't really like 'mode' on windows.
We could continue to use mode here, but really 'rights' is the right way to do
it on windows. However, without adding a whole bunch of configuration options,
there's no guaranteed right way to set the permissions on windows; it's going to
change from environment to environment. So, I guess I'll just take the easy way
out and just make the current behavior work(ish).