[COOK-1951] - add daemon_options attribute
Add the attribute with an empty array as the default value. Each of
the service config files is updated with the new options. An
additional space is added so the attribute doesn't get joined to an
existing option, but this does mean that template files will get
regenerated. A better way of handling this may be nice.