[COOK-4594] - 'Found chef-client in' log resource
Using the log resource for the 'Found chef-client in...' debug message causes it
to be reported as executed on each chef-run which I don't believe is the intention
and is confusing as it makes it appear like a resource was changed.

This will still log it when run in debug mode but not as a resource.

Signed-off-by: Sean OMeara <someara@opscode.com>