Enable changing of permissions for log directory
Previously the permissions for `node['chef_client']['log_dir'] were
hardcoded to 00750.
Sometimes it can be helpful for non-root users to view Chef logs,
and at othertimes, users may wish to set log_dir to /var/log.

In both cases it can be useful to allow the easy changing of mode
for this directory. This change permits it via an attribute.