Update Chef client config settings
This may be a bit opinionated so let me explain.

Existing wording led me to believe that setting all those config values is deprecated, period.
After inspecting the cookbook that didn't seem to be the case.

I also would like to stress that configuring environments via `client.rb` is the only realistic way
when nodes are not bootstrapped with Knife, e.g. on AWS with autoscaling groups.