Ensure upstart service picks up all config options
Using upstart service does not reliably pick up the splay and interval
options.  This commit ensures that the options are picked up, and the
chef-client daemon is reliably restarted.  Simply using 'stop
chef-client ; start chef-client' results in a non-running daemon.
Combining the trap and nohup ensures the daemon runs with the new
config.