node_name is better than fqdn
fqdn is nil and the nil behavior this replaces is horrible

node_name has to be unique or nodes will be scribbling over the
same entries in the database.  also node_name has to be not-nil or
else chef-client itself throws an exception very early, so we
can now ignore that case.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>