Systemd timeout
It can be used in the systemd timer setup to avoid stalled chef runs.
E.g. chef run gets stalled sometimes if used on the notebook and it gets
moved to another network during the run. The stalled chef process does not allow
the execution of any other chef runs.
Using the systemd timeout we can kill this stalled chef runs via systemd.

Signed-off-by: Artem Sidorenko <artem@posteo.de>