Default the systemd restart behavior to always
Digging into chef_client_updater issues it's come to our attention that
chef-client will pretty much always exit with what systemd considers to
be a clean exit code. If we term with exit code 213, the parent process
still exits with 0. Thus killing a chef run and the parent process
during an upgrade just kills the box. That's not what anyone wants, and
if they REALLY REALLY want that behavior they can set this attribute.

Signed-off-by: Tim Smith <tsmith@chef.io>