do not require start, report, or exception handlers to mange `client.rb`
While there is a rescue on `NameError` to prevent crashing when defined but not installed yet there is nothing preventing itterating over a nil value which is not valid. The original results in the following error:
```
Chef::Mixin::Template::TemplateError
    ------------------------------------
    undefined method `each' for nil:NilClass
```

After making that tweak assuming there are no handlers of any kind defined while it looks funny there is nothing wrong with what is rendered:
```
begin
rescue NameError => e
  Chef::Log.error e
end
```

Might be better to wrap the whole thing in an unless nil but that's gonna pretty ugly.

Signed-off-by: Ben Abrams <me@benabrams.it>