convert symbol-like log_location to symbols
When I set the configuration attribute log_location to ":syslog" I was getting
logging in a file named ":syslog". I discovered that Chef actually converts
"syslog" to the symbol :syslog, but that's really not documented anywhere.

I want to avoid having future discussions with anyone who sees a client.rb file
with log_location being "syslog" and thinking that it might be an error.

Signed-off-by: David Marshall <dmarshall@gmail.com>