Changes systemd timer to trigger on OnActiveSec
When running chef-client via systemd timer, the chef-client.service unit
is of Type=oneshot. oneshot services don't enter an "active" state so
depending on "onUnitActiveSec" introduces issues where the timer will not
be scheduled to run.

This changes the timer to schedule relative to the moment the timer itself
is activated[1].

[1] https://www.freedesktop.org/software/systemd/man/systemd.timer.html#Options

Signed-off-by: ngharo <ngharo@gmail.com>