Adding ability to accept upcoming Chef license
This is a backwards compatible way to prepare for the upcoming Chef
license change. Set the attribute `node['chef_client']['chef_license'] =
'accept'` (or `'accept-no-persist'` if you do not want persistence and
will always specify this attribute) to accept the Chef license. When you
upgrade to Chef Client 15 the license will be automatically accepted and
there will not be any interruption to service.

Requires running the 'chef_client::config' recipe to work correctly. If
you do not run this recipe add `chef_license "accept"` to your client
config in whatever way you manage that.

Signed-off-by: tyler-ball <tball@chef.io>