set the client_fork default to nil
this shouldn't be set in the config.rb to either true or false.

the default sets it to --fork which forces --fork to cli use
of chef-client as well, which only causes additional issues for
CLI use where the supervisor process is not necessary or
required.

the use of the --interval flag or --once or similar flags correctly
gets the default setting of fork/no-fork right and that behavior
inside the client should be left alone.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>