Prevent failures when setting certain frequency types
random_delay and frequency_modifier are not supported for all the possible frequency types. Our own tests in this cookbook were actually failing to run based on the frequency value we set. This should prevent those failures by not passing in unsupported properties. I extracted it to helper methods so I can test this easily when we move this to core.

Signed-off-by: Tim Smith <tsmith@chef.io>