add the initial chef_client_trusted_certificate resource
This is the first take at a simple resource to add trusted certificates
to chef's trusted_certificate directory. This directory is not created
out of the box so we create it if it's missing with 640 perms. For the
moment the resource just takes the cert as a string. Once we build a
file fetching mixin this should allow us to specify the content as a
string or a URI.

Signed-off-by: Tim Smith <tsmith@chef.io>