Fix functionality of append_log_file parameter for Chef 15 and below
In Chef 15 and below, using -L for sending log output also sends the output to
stdout which is not how the previous version of this cookbook worked. Currently,
all output it sent to cron email which is not wanted.

This provides a fix for users of older Chef versions and restores the previous
functionality of redirecting the output to a logfile.

Signed-off-by: Lance Albertson <lance@osuosl.org>